folder structure changes, consoleme datatable changes and eslint setup
