+++ consoleme/handlers/v2/policies.py
@@ -126,69 +126,6 @@ class PoliciesHandler(BaseAPIV2Handler):
         return
 
 
-class PoliciesTableConfigHandler(BaseHandler):
-    async def get(self):
-        """
-        /api/v2/policies_table_config
-        ---
-        get:
-            description: Retrieve Policies Table Configuration
-            responses:
-                200:
-                    description: Returns Policies Table Configuration
-        """
-        default_configuration = {
-            "expandableRows": True,
-            "tableName": "Policies",
-            "tableDescription": "View all of the AWS Resources we know about.",
-            "dataEndpoint": "/api/v2/policies?markdown=true",
-            "sortable": False,
-            "totalRows": 1000,
-            "rowsPerPage": 50,
-            "serverSideFiltering": True,
-            "columns": [
-                {
-                    "placeholder": "Account ID",
-                    "key": "account_id",
-                    "type": "input",
-                    "style": {"width": "110px"},
-                },
-                {
-                    "placeholder": "Account",
-                    "key": "account_name",
-                    "type": "input",
-                    "style": {"width": "90px"},
-                },
-                {
-                    "placeholder": "Resource",
-                    "key": "arn",
-                    "type": "input",
-                    "width": 6,
-                    "style": {"whiteSpace": "normal", "wordBreak": "break-all"},
-                },
-                {
-                    "placeholder": "Tech",
-                    "key": "technology",
-                    "type": "input",
-                    "style": {"width": "70px"},
-                },
-                {
-                    "placeholder": "Template",
-                    "key": "templated",
-                    "type": "input",
-                    "style": {"width": "100px"},
-                },
-                {"placeholder": "Errors", "key": "errors", "color": "red", "width": 1},
-            ],
-        }
-
-        table_configuration = config.get(
-            "PoliciesTableConfigHandler.configuration", default_configuration
-        )
-
-        self.write(table_configuration)
-
-
 class ManagedPoliciesHandler(BaseHandler):
     async def get(self, account_id):
         """
