+++ consoleme/handlers/v2/requests.py
@@ -658,80 +658,6 @@ class RequestDetailHandler(BaseAPIV2Handler):
         return
 
 
-class RequestsTableConfigHandler(BaseHandler):
-    async def get(self):
-        """
-        /requests_table_config
-        ---
-        get:
-            description: Retrieve Requests Table Configuration
-            responses:
-                200:
-                    description: Returns Requests Table Configuration
-        """
-        default_configuration = {
-            "expandableRows": True,
-            "tableName": "Requests",
-            "tableDescription": "View all IAM policy requests created through ConsoleMe",
-            "dataEndpoint": "/api/v2/requests?markdown=true",
-            "sortable": False,
-            "totalRows": 1000,
-            "rowsPerPage": 50,
-            "serverSideFiltering": True,
-            "columns": [
-                {
-                    "placeholder": "Username",
-                    "key": "username",
-                    "type": "input",
-                    "style": {"width": "100px"},
-                },
-                {
-                    "placeholder": "Arn",
-                    "key": "arn",
-                    "type": "input",
-                    "style": {"whiteSpace": "normal", "wordBreak": "break-all"},
-                    "width": 3,
-                },
-                {
-                    "placeholder": "Request Time",
-                    "key": "request_time",
-                    "type": "daterange",
-                },
-                {
-                    "placeholder": "Status",
-                    "key": "status",
-                    "type": "dropdown",
-                    "style": {"width": "90px"},
-                },
-                {
-                    "placeholder": "Request ID",
-                    "key": "request_id",
-                    "type": "input",
-                    "style": {"whiteSpace": "normal", "wordBreak": "break-all"},
-                    "width": 2,
-                },
-                {
-                    "placeholder": "Policy Name",
-                    "key": "policy_name",
-                    "type": "input",
-                    "style": {"width": "110px"},
-                },
-                {
-                    "placeholder": "Last Updated By",
-                    "key": "updated_by",
-                    "type": "input",
-                    "style": {"width": "110px"},
-                },
-            ],
-        }
-
-        table_configuration = config.get(
-            "RequestsTableConfigHandler.configuration", default_configuration
-        )
-
-        self.write(table_configuration)
-
-
 class RequestsWebHandler(BaseHandler):
     async def get(self):
         """
