+++ consoleme/routes.py
@@ -43,20 +43,17 @@ from consoleme.handlers.v2.generate_changes import GenerateChangesHandler
 from consoleme.handlers.v2.generate_policy import GeneratePolicyHandler
 from consoleme.handlers.v2.index import (
     EligibleRoleHandler,
-    EligibleRoleTableConfigHandler,
     FrontendHandler,
 )
 from consoleme.handlers.v2.policies import (
     ManagedPoliciesHandler,
     PoliciesHandler,
-    PoliciesTableConfigHandler,
     PolicyReviewV2Handler,
 )
 from consoleme.handlers.v2.requests import (
     RequestDetailHandler,
     RequestHandler,
     RequestsHandler,
-    RequestsTableConfigHandler,
 )
 from consoleme.handlers.v2.resources import ResourceDetailHandler
 from consoleme.handlers.v2.roles import (
@@ -123,12 +120,9 @@ def make_app(jwt_validator=None):
         (r"/api/v2/generate_policy", GeneratePolicyHandler),
         (r"/api/v2/managed_policies/(\d{12})", ManagedPoliciesHandler),
         (r"/api/v2/policies", PoliciesHandler),
-        (r"/api/v2/policies_table_config", PoliciesTableConfigHandler),
-        (r"/api/v2/role_table_config", EligibleRoleTableConfigHandler),
         (r"/api/v2/request", RequestHandler),
         (r"/api/v2/requests", RequestsHandler),
         (r"/api/v2/requests/([a-zA-Z0-9_-]+)", RequestDetailHandler),
-        (r"/api/v2/requests_table_config", RequestsTableConfigHandler),
         (r"/api/v2/roles/?", RolesHandler),
         (r"/api/v2/roles/(\d{12})", AccountRolesHandler),
         (r"/api/v2/roles/(\d{12})/(.*)", RoleDetailHandler),
