+++ ui/package.json
@@ -8,8 +8,10 @@
     "@testing-library/user-event": "^12.1.10",
     "brace": "^0.11.1",
     "js-cookie": "^2.2.1",
+    "loadash": "^1.0.0",
     "monaco-editor": "0.21.2",
     "monaco-editor-webpack-plugin": "^2.0.0",
+    "qs": "^6.9.4",
     "react": "^17.0.1",
     "react-ace": "^9.1.3",
     "react-app-rewired": "^2.1.6",
@@ -30,7 +32,9 @@
     "eject": "react-scripts eject",
     "copy:dev": "cp -a build/ ../consoleme/templates/static/ui",
     "clean:dev": "rm -rf build && rm -rf ../consoleme/templates/static/ui",
-    "build:dev": "yarn run clean:dev && yarn build && yarn run copy:dev"
+    "build:dev": "yarn run clean:dev && yarn build && yarn run copy:dev",
+    "lint": "prettier --check './src/**/*.{jsx,js}'",
+    "lint:fix": "prettier --write './src/**/*.{jsx,js}'"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -48,10 +52,15 @@
     ]
   },
   "devDependencies": {
-    "eslint-config-airbnb": "^18.2.0",
-    "eslint-plugin-prettier": "^3.1.4",
+    "eslint": "^7.13.0",
+    "eslint-config-prettier": "^6.15.0",
+    "eslint-plugin-react-hooks": "^4.2.0",
+    "nodemon": "^2.0.4",
+    "npm-run-all": "^4.1.3",
     "prettier": "^2.1.2"
   },
+  "author": "infrasec",
+  "license": "UNLICENSED",
   "homepage": "static/ui",
   "proxy": "http://localhost:8081"
 }
