+++ ui/src/components/DynamicConfig.js
@@ -1,6 +1,5 @@
-import React, { useState, useEffect } from "react";
-import { render } from "react-dom";
-import MonacoEditor from "react-monaco-editor";
+import React, { useState, useEffect } from 'react';
+import MonacoEditor from 'react-monaco-editor';
 import {
   Header,
   Button,
@@ -9,12 +8,12 @@ import {
   Segment,
   Message,
   Divider,
-} from "semantic-ui-react";
-import { sendRequestCommon } from "../helpers/utils";
+} from 'semantic-ui-react';
+import { sendRequestCommon } from '../helpers/utils';
 
 function ConsoleMeDynamicConfig() {
-  const [config, setConfig] = useState("");
-  const [configSha256, setConfigSha256] = useState("");
+  const [config, setConfig] = useState('');
+  const [configSha256, setConfigSha256] = useState('');
   const [statusMessage, setStatusMessage] = useState(null);
 
   useEffect(() => {
@@ -22,8 +21,8 @@ function ConsoleMeDynamicConfig() {
     async function fetchDynamicConfig() {
       const resJson = await sendRequestCommon(
         null,
-        "/api/v2/dynamic_config",
-        "get"
+        '/api/v2/dynamic_config',
+        'get',
       );
       setConfigSha256(resJson.sha256);
       setConfig(resJson.dynamicConfig);
@@ -35,27 +34,27 @@ function ConsoleMeDynamicConfig() {
   const updateConfig = async () => {
     const res = await sendRequestCommon(
       { new_config: config, existing_sha256: configSha256 },
-      "/api/v2/dynamic_config"
+      '/api/v2/dynamic_config',
     );
 
-    if (res.status === "success") {
+    if (res.status === 'success') {
       setStatusMessage(
         <Message positive>
           <Message.Header>Success</Message.Header>
           <Message.Content>
             Successfully updated Dynamic Configuration
           </Message.Content>
-        </Message>
+        </Message>,
       );
       setConfigSha256(res.newsha56);
       setConfig(res.newConfig);
     }
-    if (res.status === "error") {
+    if (res.status === 'error') {
       setStatusMessage(
         <Message negative>
           <Message.Header>Oops! There was a problem.</Message.Header>
           <Message.Content>{res.error}</Message.Content>
-        </Message>
+        </Message>,
       );
     }
   };
